move commands and ui elements to ticket_controller
