+++ cog/ticket/clip.py
@@ -10,7 +10,6 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from datetime import datetime, timedelta, timezone
 import time
 import re
 
@@ -36,200 +35,36 @@ class ClipTicketManagement(TicketManagement):
             for channel in self._category.channels
             if self._ticket_prefix in channel.name]
         self._embeds = self.load_embed(EMBED_PATH)
-
-    @app_commands.command(name="ticket_booth")
-    async def ticket_booth(
-        self,
-        interaction: discord.Interaction,
-    ) -> None:
-        """Sends an embed based on given parameter
-
-        Args:
-            interaction: The interaction object for the slash command
-        """
-        
-        if not self.check_user_permission(interaction.user):
-            await interaction.response.send_message(
-                "Insufficient permissions", ephemeral=True
-            )
-            return
-        
-        await interaction.response.send_modal(TicketBoothParameters(self))
     
-    async def create_ticket_booth(
-        self,
-        interaction: discord.Interaction,
-        embed_title: str,
-        embed_text: str,
-        button_label: str,
-        button_emoji: str,
-        embed_colour: int=None
-    ) -> None:
-        """Generate and send embed/button in Discord
-        
-        Args:
-            interaction: The interaction object for the slash command
-            embed_title: title of the embed
-            embed_text: description of the embed
-        """
+    def get_ticket_button(self, label, emoji):
+        return TicketButton(self, label, emoji)
 
-        embed = self.create_embed(embed_title, embed_text, embed_colour)
-        await self.send_embed(interaction.channel, embed)
-        view = discord.ui.View(timeout=None)
-        view.add_item(TicketButton(self, button_label, button_emoji))
-        try:
-           await self.send_view(interaction.channel, view)
-        except:
-            await interaction.response.send_message(
-                "ERROR: Invalid emoji, try again", ephemeral=True)
-            return
-        await interaction.response.send_message(
-            "Ticket booth created", ephemeral=True)
-    
-    @app_commands.command(
-        name="ticket_cleanup",
-        description="deletes all tickets older than 2 weeks"
-    )
-    async def clean_tickets(self, interaction) -> None:
-        """Delete all tickets with the last message sent before the stale time.
-        Note this method uses channel.history not channel.last_message as
-        channel.last_message may point to a deleted message which throws an 
-        error
-        
-        Args:
-            interaction: The interaction object for the slash command
-        """
-        
-        # check user permissions
-        if not self.check_user_permission(interaction.user):
-            await interaction.response.send_message(
-                "Insufficient permissions", ephemeral=True
-            )
-            return
-        
-        present = datetime.now(timezone.utc)
-        stale_date = present - self._time_until_ticket_stale
-        tickets_deleted = 0
-        
-        await interaction.response.defer(thinking=True, ephemeral=True)
-        
-        for channel in self._category.channels:
-            last_message = channel.history(limit=1)
-            date = [message.created_at async for message in last_message][0]
-            if (date < stale_date):
-                await channel.delete()
-                tickets_deleted += 1
-            
-        await interaction.followup.send(
-            f"{tickets_deleted} ticket(s) deleted")
-        
-        
 class TicketButton(
     discord.ui.DynamicItem[discord.ui.Button],
     template=r'make_ticket:([0-9]+)'):
-    """Dynamic, persistent button to handle ticket creation"""
-    
-    def __init__(self, ticket_manager, label=None, emoji=None):
-        self._ticket_manager = ticket_manager
-        super().__init__(
-            discord.ui.Button(
-                label=label,
-                emoji=emoji,
-                style=discord.ButtonStyle.blurple,
-                custom_id=f"make_ticket:{int(time.time())}"
+        """Dynamic, persistent button to handle ticket creation"""
+        
+        def __init__(self, ticket_manager, label=None, emoji=None):
+            self._ticket_manager = ticket_manager
+            super().__init__(
+                discord.ui.Button(
+                    label=label,
+                    emoji=emoji,
+                    style=discord.ButtonStyle.blurple,
+                    custom_id=f"make_ticket:{int(time.time())}"
+                )
             )
-        )
 
-    @classmethod
-    async def from_custom_id(cls,
-                             interaction: discord.Interaction,
-                             item: discord.ui.Button, 
-                             match: re.Match[str]):
-        time = int(match.group(1))
-        ticket_manager = interaction.client.ticket_manager_instance
-        return cls(ticket_manager)
+        @classmethod
+        async def from_custom_id(cls,
+                                interaction: discord.Interaction,
+                                item: discord.ui.Button, 
+                                match: re.Match[str]):
+            ticket_manager = interaction.client.instances[TICKET_PREFIX]
+            return cls(ticket_manager)
 
-    async def callback(self, interaction):
-        await self._ticket_manager.create_ticket(interaction)
-        
-    
-class TicketBoothParameters(discord.ui.Modal):
-    """Set parameters for ticket booth here"""
-    
-    def __init__(
-        self, 
-        ticket_manager: ClipTicketManagement
-    ) -> None:
-
-        super().__init__()
-        self._ticket_manager = ticket_manager
-        
-    # Questions in form
-    title = "Configure ticket booth"
-    embed_title = discord.ui.TextInput(
-        style=discord.TextStyle.short,
-        label="Title", 
-        placeholder="Name of the ticket"
-    )
-    embed_text = discord.ui.TextInput(
-        style=discord.TextStyle.long,
-        label="Description", 
-        placeholder="What is this ticket for?"
-    )
-    button_title=discord.ui.TextInput(
-        style=discord.TextStyle.short,
-        required=True,
-        label="Button Title", 
-        placeholder="Text on button"
-    )
-    button_emoji=discord.ui.TextInput(
-        style=discord.TextStyle.short,
-        required=False,
-        default=None,
-        label="Button Emoji", 
-        placeholder="Emoji on button"
-    )
-    embed_colour = discord.ui.TextInput(
-        style=discord.TextStyle.short,
-        required=False,
-        default=None,
-        label="Colour",
-        placeholder="Hex colour code (starts with # or 0x)"
-    )
-    async def on_submit(self, interaction: discord.Interaction):
-        
-
-        emoji = self.button_emoji.value
-        if not emoji:
-            emoji = None
-            
-        # handle colour code
-        colour = self.embed_colour.value
-        colour = colour.lstrip("#")
-        colour = colour.lstrip("0x")
-        
-        if not colour:
-            colour = None
-        else:
-            if len(colour) != 6:
-                await interaction.response.send_message(
-                    "Hexcode must have 6 characters", ephemeral=True)
-            else:     
-                try:
-                    colour = int(colour, 16)
-                except:
-                    await interaction.response.send_message(
-                        "Hexcode not valid", ephemeral=True)
-                    return
-        
-        await self._ticket_manager.create_ticket_booth(
-                interaction,
-                self.embed_title.value,
-                self.embed_text.value,
-                self.button_title.value,
-                emoji,
-                colour,
-            )
+        async def callback(self, interaction):
+            await self._ticket_manager.create_ticket(interaction)
 
 async def setup(bot: commands.Bot) -> None:
     """A hook for the bot to register the ClipTicketManagement cog.
@@ -241,6 +76,7 @@ async def setup(bot: commands.Bot) -> None:
     """
     instance = ClipTicketManagement(bot)
     # store instance of ClipTicketManagement to be used in TicketButton
-    bot.ticket_manager_instance = instance 
+
+    bot.instances[TICKET_PREFIX] = instance
     bot.add_dynamic_items(TicketButton)
     await bot.add_cog(instance)
